+++ sdk/serialization/azure-json/src/main/java/module-info.java
@@ -3,7 +3,7 @@
 
 module com.azure.json {
     exports com.azure.json;
-    exports com.azure.json.tree;
+    exports com.azure.json.models;
 
     uses com.azure.json.JsonProvider;
 }
