+++ sdk/serialization/azure-json/src/test/java/com/azure/json/contract/models/JsonBooleanContractTests.java
@@ -1,14 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.json.contract.tree;
+package com.azure.json.contract.models;
 
 import com.azure.json.JsonOptions;
 import com.azure.json.JsonProvider;
 import com.azure.json.JsonReader;
 import com.azure.json.JsonWriter;
 import com.azure.json.implementation.StringBuilderWriter;
-import com.azure.json.tree.JsonElement;
-import com.azure.json.tree.JsonNull;
+import com.azure.json.models.JsonBoolean;
+import com.azure.json.models.JsonElement;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -22,12 +22,12 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
- * Tests the contract of {@link JsonNull}.
+ * Tests the contract of {@link JsonBoolean}.
  * <p>
  * All implementations of {@link JsonProvider} must create a subclass of this test class and pass all tests as they're
  * written to be considered an acceptable implementation.
  */
-public abstract class JsonNullContractTests {
+public abstract class JsonBooleanContractTests {
     /**
      * Creates an instance {@link JsonProvider} that will be used by a test.
      *
@@ -37,40 +37,48 @@ public abstract class JsonNullContractTests {
 
     @Test
     public void kindCheck() {
-        JsonElement element = JsonNull.getInstance();
-        assertTrue(element.isNull());
+        JsonElement element = JsonBoolean.getInstance(true);
+        assertTrue(element.isBoolean());
         assertFalse(element.isArray());
         assertFalse(element.isObject());
         assertFalse(element.isString());
         assertFalse(element.isNumber());
-        assertFalse(element.isBoolean());
+        assertFalse(element.isNull());
     }
 
-    @Test
-    public void fromJson() throws IOException {
-        try (JsonReader reader = getJsonProvider().createReader("null", new JsonOptions())) {
-            JsonNull jsonNull = JsonNull.fromJson(reader);
-            assertEquals("null", jsonNull.toJsonString());
-            assertSame(JsonNull.getInstance(), jsonNull);
-        }
-    }
-
-    @Test
-    public void toJson() throws IOException {
-        JsonNull jsonNull = JsonNull.getInstance();
-        try (StringBuilderWriter writer = new StringBuilderWriter()) {
-            try (JsonWriter jsonWriter = getJsonProvider().createWriter(writer, new JsonOptions())) {
-                jsonNull.toJson(jsonWriter);
+    @ParameterizedTest
+    @ValueSource(strings = { "true", "false" })
+    public void fromJson(String json) throws IOException {
+        try (JsonReader reader = getJsonProvider().createReader(json, new JsonOptions())) {
+            JsonBoolean jsonBoolean = JsonBoolean.fromJson(reader);
+            assertEquals(json, jsonBoolean.toJsonString());
+            if ("true".equals(json)) {
+                assertTrue(jsonBoolean.getValue());
+                assertSame(JsonBoolean.getInstance(true), jsonBoolean);
+            } else {
+                assertFalse(jsonBoolean.getValue());
+                assertSame(JsonBoolean.getInstance(false), jsonBoolean);
             }
-            assertEquals("null", writer.toString());
         }
     }
 
     @ParameterizedTest
-    @ValueSource(strings = { "true", "10", "10.0", "\"hello\"", "[]", "{}" })
+    @ValueSource(booleans = { true, false })
+    public void toJson(boolean value) throws IOException {
+        JsonBoolean jsonBoolean = JsonBoolean.getInstance(value);
+        try (StringBuilderWriter writer = new StringBuilderWriter()) {
+            try (JsonWriter jsonWriter = getJsonProvider().createWriter(writer, new JsonOptions())) {
+                jsonBoolean.toJson(jsonWriter);
+            }
+            assertEquals(Boolean.toString(value), writer.toString());
+        }
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = { "null", "10", "10.0", "\"hello\"", "[]", "{}" })
     public void invalidFromJsonStartingPoints(String json) throws IOException {
         try (JsonReader reader = getJsonProvider().createReader(json, new JsonOptions())) {
-            assertThrows(IllegalStateException.class, () -> JsonNull.fromJson(reader));
+            assertThrows(IllegalStateException.class, () -> JsonBoolean.fromJson(reader));
         }
     }
 }
