+++ sdk/serialization/azure-json/src/test/java/com/azure/json/implementation/models/DefaultJsonArrayContractTests.java
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.json.implementation.tree;
+package com.azure.json.implementation.models;
 
 import com.azure.json.JsonProvider;
-import com.azure.json.contract.tree.JsonArrayContractTests;
+import com.azure.json.contract.models.JsonArrayContractTests;
 import com.azure.json.implementation.DefaultJsonProvider;
 
 public class DefaultJsonArrayContractTests extends JsonArrayContractTests {
