+++ sdk/serialization/azure-json/src/test/java/com/azure/json/implementation/models/DefaultJsonBooleanContractTests.java
@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.json.implementation.tree;
+package com.azure.json.implementation.models;
 
 import com.azure.json.JsonProvider;
-import com.azure.json.contract.tree.JsonElementContractTests;
+import com.azure.json.contract.models.JsonBooleanContractTests;
 import com.azure.json.implementation.DefaultJsonProvider;
 
-public class DefaultJsonElementContractTests extends JsonElementContractTests {
+public class DefaultJsonBooleanContractTests extends JsonBooleanContractTests {
     @Override
     protected JsonProvider getJsonProvider() {
         return new DefaultJsonProvider();
