+++ sdk/serialization/azure-json/src/test/java/com/azure/json/implementation/models/DefaultJsonNullContractTests.java
@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.json.implementation.tree;
+package com.azure.json.implementation.models;
 
 import com.azure.json.JsonProvider;
-import com.azure.json.contract.tree.JsonStringContractTests;
+import com.azure.json.contract.models.JsonNullContractTests;
 import com.azure.json.implementation.DefaultJsonProvider;
 
-public class DefaultJsonStringContractTests extends JsonStringContractTests {
+public class DefaultJsonNullContractTests extends JsonNullContractTests {
     @Override
     protected JsonProvider getJsonProvider() {
         return new DefaultJsonProvider();
