+++ sdk/serialization/azure-json/src/test/java/com/azure/json/implementation/models/DefaultJsonNumberContractTests.java
@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.json.implementation.tree;
+package com.azure.json.implementation.models;
 
 import com.azure.json.JsonProvider;
-import com.azure.json.contract.tree.JsonBooleanContractTests;
+import com.azure.json.contract.models.JsonNumberContractTests;
 import com.azure.json.implementation.DefaultJsonProvider;
 
-public class DefaultJsonBooleanContractTests extends JsonBooleanContractTests {
+public class DefaultJsonNumberContractTests extends JsonNumberContractTests {
     @Override
     protected JsonProvider getJsonProvider() {
         return new DefaultJsonProvider();
